Move compiled scripts to separate build directory
This makes it easier to manage as we can simplify the ignore list, the
cleaning pattern, and what files need to be compiled. Note, the make
recipe will not compile files in src that end in either .sh or begin
with an underscore. The latter is not needed now but might be handy in
the future.