Switch to using @ instead of * for accesing array
There was a bug that manifested itself when a search term had more than
one word. E.g., "hand in glove" was treated as `cmd hand in glove` so
the "glove" was considered as the directory. The issue was that using *
for accessing the parsed arguments would not properly treat whitespace.
Using @ will not expand the words in the arguments.