Fix bug of limiting results and 2nd search term
Before, we used grep to apply the second search term and that was
applied after limiting the results. As such, if we wanted to both use a
second search term and limit the results, we need to be lucky that
limiting the results didn't filter out the songs we were looking for.
Limiting the results should happen after apply the second search term.