Run tests in GitLab CI
Just one job that takes care of both building and running the tests.
I had wanted to use multiple stages/jobs (whatever is the right
terminology) but had trouble given the indepdence of the stages/jobs.
For example, fd-find was installed in the build stage but was not
available in the testing stage. I can revisit using multiple
stages/jobs when I want to deploy the built scripts.